refactor(chat): give tool replay reconciliation its own module
